Update wave-dependent growth of new ice
Add check on the value of wave peak frequency being too small in the calculation of new ice floe size in the presence of waves, in subroutine ice_wav_newice of module icewav.F90 (this should not affect behaviour as if wave peak frequency is zero then presumably this indicates missing data or the absence of waves, for which significant wave height will also be zero and is already checked for. However adding this check is safer and ensures the namelist default new ice floe size is used consistently).