Fix bug in heat/salt content update during lateral growth
Re-structure and correct the conditions/order in the thickness-category loop updating new ice heat and salt content in subroutine ice_thd_do.

The original loop is adapted in commit #831c4d2 to account for lateral growth of existing ice in all thickness categories when FSD is activated (ln_fsd = T) as well as open-water growth. But the condition checking that there is pre-existing ice in the category jcat to which the latter is added (which affects the updated heat/salt calculation) is erronously included for all loop indices rather than just jcat. This means that all categories are always updated with the new heat/salinity value, whereas only those categories to which new ice is added should be updated. This is just jcat if ln_fsd = F; if ln_fsd = T, all categories are supposed to be updated, but only if there is ice to begin with and this is not checked.

This commit fixes the above by retaining the outer loop over thickness categories, but changing the loop steps to (1) calculate total new ice volume from lateral growth of existing ice and open-water growth of new ice, (2) check if the total new ice volume is non-zero, and if so (3) update the category heat/salinity using the original code that first checks whether there is pre-existing ice in said category and then using the appropriate calculation with the total new ice volume.

This fix reproduces the original (pre-FSD modifications) loop behaviour, and test simulation, when ln_fsd = F, and corrects it when ln_fsd = T.